Split out routing logic to its own class
